requirement.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
numpy==1.23.5
opencv-python==4.6.0.66
python-dotenv==0.21.0
requests==2.28.1
setuptools==60.2.0
urllib3==1.26.13
wheel==0.37.1
yarl==1.8.2
python-bsonjs==0.3.0
pillow==9.3.0